refactor: JPA Repository를 domain에서 infrastructure로 이동 (#21)
- command/domain/repository/ 디렉터리 삭제
- ChannelOrderRepository, EasypostShipmentInvoiceRepository, ChannelApiRepository를
  command/infrastructure/repository/로 이동 (DB 접근은 infrastructure에서만)
- 영향받는 서비스 및 테스트 파일 import 일괄 수정
