+++ src/main/java/com/conk/integration/command/infrastructure/repository/EasypostShipmentInvoiceRepository.java
@@ -1,4 +1,4 @@
-package com.conk.integration.command.domain.repository;
+package com.conk.integration.command.infrastructure.repository;
 
 import com.conk.integration.command.domain.aggregate.EasypostShipmentInvoice;
 import org.springframework.data.jpa.repository.JpaRepository;
