+++ src/test/java/com/conk/integration/command/application/service/ShopifyOrderSyncServiceTest.java
@@ -4,7 +4,7 @@ import com.conk.integration.command.application.dto.response.ShopifyOrderRespons
 import com.conk.integration.command.application.service.shopify.ShopifyOrderSyncService;
 import com.conk.integration.command.domain.aggregate.ChannelOrder;
 import com.conk.integration.command.domain.aggregate.enums.OrderChannel;
-import com.conk.integration.command.domain.repository.ChannelOrderRepository;
+import com.conk.integration.command.infrastructure.repository.ChannelOrderRepository;
 import com.conk.integration.command.infrastructure.service.ShopifyOrderClient;
 import com.conk.integration.query.dto.ShopifyCredentialDto;
 import com.conk.integration.query.service.ChannelApiQueryService;
