+++ src/test/java/com/conk/integration/e2e/flow/IntegrationTest.java
@@ -4,7 +4,7 @@ import com.conk.integration.command.domain.aggregate.*;
 import com.conk.integration.command.domain.aggregate.embeddable.ChannelApiId;
 import com.conk.integration.command.domain.aggregate.enums.CarrierType;
 import com.conk.integration.command.domain.aggregate.enums.OrderChannel;
-import com.conk.integration.command.domain.repository.*;
+import com.conk.integration.command.infrastructure.repository.*;
 import com.conk.integration.query.mapper.ChannelApiMapper;
 import com.conk.integration.command.application.service.ChannelFulfillmentDispatchService;
 import com.conk.integration.command.application.service.shopify.ShopifyOrderSyncService;
