+++ src/test/java/com/conk/integration/query/mapper/ChannelOrderInvoiceMapperTest.java
@@ -2,7 +2,7 @@ package com.conk.integration.query.mapper;
 
 import com.conk.integration.command.domain.aggregate.ChannelOrder;
 import com.conk.integration.command.domain.aggregate.enums.OrderChannel;
-import com.conk.integration.command.domain.repository.ChannelOrderRepository;
+import com.conk.integration.command.infrastructure.repository.ChannelOrderRepository;
 import com.conk.integration.query.dto.InvoiceTargetDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
